Move language picker to shared components
